LayoutTests/platform/ios-simulator/editing/pasteboard/paste-wrapped-blockquote-into-nonblockquote-expected.txt

This test ensures the copied the newline is NOT inside the blockquote -- a '<div><div><br></div></div>' should be appended as the next sibling of the blockquote. 

| <div>
|   id="copyDiv"
|   <blockquote>
|     type="cite"
|     <div>
|       <font>
|         color="#000000"
|         <span>
|           style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|           "line 1"
|     <div>
|       <font>
|         color="#000000"
|         <span>
|           style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|           "line 2"
|   <div>
|     <div>
|       <#selection-caret>
|       <br>
